+++ R/xcorr.R
@@ -1,29 +1,21 @@
 #' Spectrogram cross-correlation 
 #' 
 #' \code{xcorr} estimates the similarity of two spectrograms by means of cross-correlation
-#' @usage xcorr(X, wl =512, frange= NULL, ovlp=90, dens=0.9, bp= NULL, wn='hanning', 
+#' @usage xcorr(X, wl =512, bp = NULL, ovlp=90, dens=0.9, wn='hanning', 
 #' cor.method = "pearson", parallel = 1, path = NULL, pb = TRUE, na.rm = FALSE,
-#'  dfrange = FALSE, cor.mat = TRUE)
+#'  cor.mat = TRUE)
 #' @param  X 'selection_table', 'extended_selection_table' or data frame containing columns for sound files (sound.files), 
 #' selection number (selec), and start and end time of signal (start and end).
 #' @param wl A numeric vector of length 1 specifying the window length of the spectrogram, default 
 #' is 512.
-#' @param frange A numeric vector of length 2 setting the upper and lower frequency limits (in kHz) 
+#' @param bp A numeric vector of length 2 setting the upper and lower frequency limits (in kHz) 
 #' in which to compare the signals or "frange" (default) to indicate that values in 'bottom.freq'
-#' and 'top.freq' columns will be used as frequency limits. Alternatively, the \code{\link{dfts}} function can 
-#' be used to determine this parameter if \code{dfrange = TRUE} and \code{frange = NULL}. This method is more 
-#' adequate for pure tone signals. Default is \code{NULL}. Either 'frange' should be provided or 
-#' set \code{dfrange = TRUE}.
-#' @param dfrange Logical. If \code{TRUE} the \code{\link{dfts}} function can is used to determine the frequency range in which to compare signals.
-#' Ignored if 'frange' is provided. 
+#' and 'top.freq' columns will be used as frequency limits. Default is \code{NULL}.
 #' @param ovlp Numeric vector of length 1 specifying \% of overlap between two 
 #' consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 90. High values of ovlp 
 #' slow down the function but produce more accurate results.
 #' @param dens Numeric vector of length 1 specifying the approximate density of points in which to sample amplitude. 
 #' See \code{\link[monitoR]{makeTemplate}}. Deafult is 0.9.
-#' @param bp A numeric vector of length 2 for the lower and upper limits of a 
-#' frequency bandpass filter (in kHz) in which to detect dominant frequency. 
-#' Only applied when frange is \code{NULL}. Default is \code{NULL}.
 #' @param wn A character vector of length 1 specifying the window name as in \code{\link[seewave]{ftwindow}}. 
 #' @param cor.method A character vector of length 1 specifying the correlation method as in \code{\link[stats]{cor}}.
 #' @param parallel Numeric. Controls whether parallel computing is applied.
@@ -61,7 +53,7 @@
 #' writeWave(Phae.long3, "Phae.long3.wav")
 #' writeWave(Phae.long4, "Phae.long4.wav")
 #'
-#'xcor <- xcorr(X = selec.table, wl = 300, frange = c(2, 9), ovlp = 90,
+#'xcor <- xcorr(X = selec.table, wl = 300, bp = c(2, 9), ovlp = 90,
 #'dens = 1, wn = 'hanning', cor.method = "pearson")
 #' 
 #' }
@@ -75,9 +67,9 @@
 #' }
 # last modification on may-7-2018 (MAS)
 
-xcorr <- function(X = NULL, wl = 512, frange = NULL, ovlp = 90, dens = 0.9, bp = NULL, wn ='hanning', 
-                  cor.method = "pearson", parallel = 1, path = NULL,
-                  pb = TRUE, na.rm = FALSE, dfrange = FALSE, cor.mat = TRUE)
+xcorr <- function(X = NULL, wl = 512, bp = NULL, ovlp = 90, dens = 0.9, 
+                  wn ='hanning', cor.method = "pearson", parallel = 1, 
+                  path = NULL, pb = TRUE, na.rm = FALSE, cor.mat = TRUE)
 {
   
   # reset working directory 
@@ -125,21 +117,17 @@ xcorr <- function(X = NULL, wl = 512, frange = NULL, ovlp = 90, dens = 0.9, bp =
   #stop if only 1 selection
   if (nrow(X) == 1) stop("you need more than one selection to do cross-correlation")
   
-  #if bp is not vector or length!=2 stop
-  if (!is.null(bp)) {if (!is.vector(bp)) stop("'bp' must be a numeric vector of length 2") else{
-    if (!length(bp) == 2) stop("'bp' must be a numeric vector of length 2")}}
   
-  #if flim is not vector or length!=2 stop
-  if (frange[1] != "frange")
-  {
-    if (is.null(frange) & !dfrange) stop("either 'frange' must be provided or 'dfrange' set to TRUE")
-  if (!is.null(frange) & !is.vector(frange)) stop("'frange' must be a numeric vector of length 2 or set to 'frange'") else
-      if (!is.null(frange) & !length(frange) == 2) stop("'frange' must be a numeric vector of length 2 or set to 'frange'")
-    } else  {if (!any(names(X) == "bottom.freq") & !any(names(X) == "top.freq")) stop("'frange' = frange requires bottom.freq and top.freq columns in X")
-      if (any(is.na(c(X$bottom.freq, X$top.freq)))) stop("NAs found in bottom.freq and/or top.freq") 
-      if (any(c(X$bottom.freq, X$top.freq) < 0)) stop("Negative values found in bottom.freq and/or top.freq") 
-      if (any(X$top.freq - X$bottom.freq < 0)) stop("top.freq should be higher than bottom.freq")
-    }
+  # bp checking
+  if (bp[1] != "frange")
+  {if (!is.vector(bp)) stop("'bp' must be a numeric vector of length 2") else{
+    if (!length(bp) == 2) stop("'bp' must be a numeric vector of length 2")} 
+  } else
+  {if (!any(names(X) == "bottom.freq") & !any(names(X) == "top.freq")) stop("'bp' = frange requires bottom.freq and top.freq columns in X")
+    if (any(is.na(c(X$bottom.freq, X$top.freq)))) stop("NAs found in bottom.freq and/or top.freq") 
+    if (any(c(X$bottom.freq, X$top.freq) < 0)) stop("Negative values found in bottom.freq and/or top.freq") 
+    if (any(X$top.freq - X$bottom.freq < 0)) stop("top.freq should be higher than low.f")
+  }
   
   #if wl is not vector or length!=1 stop
   if (!is.numeric(wl)) stop("'wl' must be a numeric vector of length 1") else {
@@ -172,13 +160,9 @@ xcorr <- function(X = NULL, wl = 512, frange = NULL, ovlp = 90, dens = 0.9, bp =
     }
   }
   
-  
-# if frange was not provided the range is calculated with dominant frequency range  
-if (dfrange & is.null(frange)) {df <- dfts(X, wl =300, img = FALSE, length.out = 50, parallel = parallel, clip.edges = TRUE)
-  df <- df[, 3:ncol(df)]
-frq.lim = c(min(df, na.rm = TRUE), max(df, na.rm = TRUE))
-} else 
-  if (frange == "frange") frq.lim <- c(min(X$bottom.freq), max(X$top.freq)) else frq.lim <- frange
+  # if bp is frange
+  if (bp[1] == "frange") frq.lim <- c(min(X$bottom.freq), max(X$top.freq)) 
+  else frq.lim <- bp
 
   # If parallel is not numeric
   if (!is.numeric(parallel)) stop("'parallel' must be a numeric vector of length 1")